IsAdmin updates + orderDetail

+++ src/app/api/categories/route.js
@@ -4,14 +4,18 @@ import { Category } from "../../../models/category";
 export async function POST(req) {
     mongoose.connect(process.env.MONGO_URL);
     const {name} = await req.json();
-    const categoryDoc = await Category.create({name});
-    return Response.json(categoryDoc);
+    if (await isAdmin()) {
+        const categoryDoc = await Category.create({name});
+        return Response.json(categoryDoc);
+    } 
 }
 
 export async function PUT(req) {
     mongoose.connect(process.env.MONGO_URL);
     const {_id, name} = await req.json();
-    await Category.updateOne({_id}, {name});
+    if (await isAdmin()) {
+        await Category.updateOne({_id}, {name});
+    }
     return Response.json(true);
 }
 
@@ -26,6 +30,8 @@ export async function DELETE(req) {
     mongoose.connect(process.env.MONGO_URL);
     const url = new URL(req.url);
     const _id = url.searchParams.get('_id');
-    await Category.deleteOne({_id});
+    if (await isAdmin()) {
+        await Category.deleteOne({_id});
+    }
     return Response.json(true);
 }
